Use an int to store the 'base' value, when decoding a special debug opcode, to prevent an incorrect negative value
git-svn-id: https://smali.googlecode.com/svn/trunk@508 55b6fa8a-2a1e-11de-a435-ffa8d773f76a

+++ dexlib/src/main/java/org/jf/dexlib/Debug/DebugInstructionIterator.java
@@ -140,7 +140,7 @@ public class DebugInstructionIterator {
                 }
                 default:
                 {
-                    byte base = (byte)((debugOpcode & 0xFF) - 0x0A);
+                    int base = ((debugOpcode & 0xFF) - 0x0A);
                     processDebugInstruction.ProcessSpecialOpcode(startOffset, debugOpcode, (base % 15) - 4, base / 15);
                 }
             }
